requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# AI/ML
google-generativeai==0.7.2

# Web Scraping (from existing final.py)
cloudscraper==1.2.71
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development (optional)
python-dotenv[cli]==1.0.0